@@ -13,21 +13,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.6",
-    "@angular/cdk": "^13.2.6",
-    "@angular/common": "~13.2.6",
-    "@angular/compiler": "~13.2.6",
-    "@angular/core": "~13.2.6",
-    "@angular/forms": "~13.2.6",
-    "@angular/localize": "~13.2.6",
-    "@angular/material": "^13.2.6",
-    "@angular/platform-browser": "~13.2.6",
-    "@angular/platform-browser-dynamic": "~13.2.6",
-    "@angular/router": "~13.2.6",
-    "@angular/service-worker": "~13.2.6",
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/localize": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@angular/service-worker": "^14.3.0",
     "chart.js": "^3.7.0",
     "date-fns": "^2.28.0",
-    "dexie": "^3.2.0",
+    "dexie": "^3.2.3",
     "ng2-charts": "^3.0.0-rc.7",
     "qrious": "^4.0.2",
     "rxjs": "~7.4.0",
@@ -36,27 +36,27 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.6",
-    "@angular-eslint/builder": "13.1.0",
-    "@angular-eslint/eslint-plugin": "13.1.0",
-    "@angular-eslint/eslint-plugin-template": "13.1.0",
-    "@angular-eslint/schematics": "13.1.0",
-    "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "~13.2.6",
-    "@angular/compiler-cli": "~13.2.6",
+    "@angular-devkit/build-angular": "^14.2.11",
+    "@angular-eslint/builder": "14.4.0",
+    "@angular-eslint/eslint-plugin": "14.4.0",
+    "@angular-eslint/eslint-plugin-template": "14.4.0",
+    "@angular-eslint/schematics": "14.4.0",
+    "@angular-eslint/template-parser": "14.4.0",
+    "@angular/cli": "^14.2.11",
+    "@angular/compiler-cli": "^14.3.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
-    "eslint": "^8.2.0",
+    "@typescript-eslint/eslint-plugin": "^5.36.2",
+    "@typescript-eslint/parser": "^5.36.2",
+    "eslint": "^8.23.0",
     "jasmine-core": "~3.10.0",
-    "karma": "~6.3.0",
+    "karma": "~6.4.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
     "source-map-explorer": "^2.5.2",
-    "typescript": "~4.5.2"
+    "typescript": "~4.8.4"
   }
 }
