@@ -1,5 +1,5 @@
 import { Injectable, OnDestroy } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { Subject, takeUntil } from 'rxjs';
 import { feeConfig } from '../data/fee-config';
 
@@ -7,16 +7,16 @@ import { feeConfig } from '../data/fee-config';
 export class FeeFormBuilder implements OnDestroy {
   private destroy = new Subject<void>();
 
-  buildForm(): FormGroup {
-    const form = new FormGroup({
-      uuid: new FormControl(null),
-      houseUuid: new FormControl(null, {validators: [Validators.required]}),
-      type: new FormControl(null, {validators: [Validators.required]}),
-      quantity: new FormControl(0, {validators: [Validators.required, Validators.min(1)]}),
-      unit: new FormControl(null, {validators: [Validators.required]}),
-      startAt: new FormControl(null, {validators: [Validators.required]}),
-      endAt: new FormControl(null, {validators: [Validators.required]}),
-      value: new FormControl(null, {validators: [Validators.required, Validators.min(0.01)]}),
+  buildForm(): UntypedFormGroup {
+    const form = new UntypedFormGroup({
+      uuid: new UntypedFormControl(null),
+      houseUuid: new UntypedFormControl(null, {validators: [Validators.required]}),
+      type: new UntypedFormControl(null, {validators: [Validators.required]}),
+      quantity: new UntypedFormControl(0, {validators: [Validators.required, Validators.min(1)]}),
+      unit: new UntypedFormControl(null, {validators: [Validators.required]}),
+      startAt: new UntypedFormControl(null, {validators: [Validators.required]}),
+      endAt: new UntypedFormControl(null, {validators: [Validators.required]}),
+      value: new UntypedFormControl(null, {validators: [Validators.required, Validators.min(0.01)]}),
     });
 
     form.get('type')!.valueChanges
