@@ -1,12 +1,12 @@
 import { Injectable } from '@angular/core';
-import { FormControl, FormGroup } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup } from '@angular/forms';
 
 @Injectable()
 export class HouseFormBuilder {
-  buildForm(): FormGroup {
-    return new FormGroup({
-      uuid: new FormControl(null),
-      name: new FormControl(null)
+  buildForm(): UntypedFormGroup {
+    return new UntypedFormGroup({
+      uuid: new UntypedFormControl(null),
+      name: new UntypedFormControl(null)
     })
   }
 }
