@@ -1,13 +1,13 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { AbstractControl, FormControl } from '@angular/forms';
+import { AbstractControl, UntypedFormControl } from '@angular/forms';
 
 @Pipe({
   name: 'asFormControl'
 })
 export class AsFormControlPipe implements PipeTransform {
 
-  transform(control: AbstractControl | null): FormControl {
-    return control as FormControl;
+  transform(control: AbstractControl | null): UntypedFormControl {
+    return control as UntypedFormControl;
   }
 
 }
